requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
addict=2.4.0=pypi_0
asttokens=2.0.8=pypi_0
attr=0.3.2=pypi_0
attrs=22.1.0=pypi_0
backcall=0.2.0=pypi_0
backpack=0.1=pypi_0
blas=1.0=mkl
ca-certificates=2022.07.19=h06a4308_0
cachecontrol=0.12.11=pypi_0
cachetools=5.2.0=pypi_0
certifi=2022.9.24=py38h06a4308_0
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
codecov=2.1.12=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.0.5=pypi_0
coverage=6.5.0=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.11.0=pypi_0
cython=0.29.32=pypi_0
decorator=5.1.1=pypi_0
exceptiongroup=1.0.0=pypi_0
executing=1.1.1=pypi_0
flake8=5.0.4=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
future=0.18.2=pypi_0
giflib=5.2.1=h7b6447c_0
google-auth=2.14.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
gpytorch=1.4.2=pypi_0
grpcio=1.50.0=pypi_0
h5py=3.7.0=pypi_0
hdmedians=0.14.2=pypi_0
idna=3.4=pypi_0
imageio=2.23.0=pypi_0
importlib-metadata=5.0.0=pypi_0
iniconfig=1.1.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
interrogate=1.5.0=pypi_0
ipython=8.5.0=pypi_0
isort=4.3.21=pypi_0
jedi=0.18.1=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.4=pypi_0
kmeans1d=0.3.1=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
linear-operator=0.2.0=pypi_0
lockfile=0.12.2=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.6.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mccabe=0.7.0=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mmcls=0.24.0=pypi_0
mmcv-full=1.2.7=pypi_0
msgpack=1.0.4=pypi_0
natsort=8.2.0=pypi_0
ncurses=6.3=h5eee18b_3
networkx=2.8.8=pypi_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.23.3=py38h14f4228_0
numpy-base=1.23.3=py38h31eccc5_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.5.1.48=pypi_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.5.1=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=py38hace64e9_1
pip=22.2.2=py38h06a4308_0
pluggy=1.0.0=pypi_0
prettytable=3.4.1=pypi_0
progress=1.6=pypi_0
prompt-toolkit=3.0.31=pypi_0
protobuf=3.19.6=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py=1.11.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycodestyle=2.9.1=pypi_0
pydevd-pycharm=222.4345.23=pypi_0
pyflakes=2.5.0=pypi_0
pygments=2.13.0=pypi_0
pyparsing=3.0.9=pypi_0
pytest=7.2.0=pypi_0
python=3.8.13=haa1d7c7_1
python-dateutil=2.8.2=pypi_0
pytz=2022.6=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
sam-pytorch=0.0.1=pypi_0
scikit-bio=0.5.6=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.2.1=pypi_0
scipy=1.9.3=pypi_0
seaborn=0.12.1=pypi_0
setuptools=63.4.1=py38h06a4308_0
simplejson=3.17.6=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.3=h5082296_0
stack-data=0.5.1=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.10.10=pypi_0
timm=0.3.2=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
torch=1.7.1+cu110=pypi_0
torchcontrib=0.0.2=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.8.2+cu110=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.5.0=pypi_0
typing-extensions=4.4.0=pypi_0
urllib3=1.26.12=pypi_0
wcwidth=0.2.5=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xdoctest=1.1.0=pypi_0
xz=5.2.6=h5eee18b_0
yapf=0.32.0=pypi_0
zipp=3.10.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
